Add first validation of ResponseXML via XPath
To test this try the following:
Von: 8011160 (Berlin Hbf)
Nach: 8002549 (Hamburg Hbf)
Datum: 2016-10-05
Uhrzeit: 16:00
Erwachsene: any
Klasse: 2 (for valid) and 1 (for invalid)

However, this does not seem to match the current DB OnlineTicket Service. It does match the XMLs though, so it's (hopefully)  not a bug on the validation method.